projects/defensive/incident_response_engine/requirements.txt
pyyaml>=6.0
rich>=13.0
jinja2>=3.1
